src/requirements.txt
attrs==22.2.0
click==8.1.3
exceptiongroup==1.1.1
gunicorn==20.1.0
h11==0.14.0
iniconfig==2.0.0
loguru==0.6.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.2
python-dotenv==1.0.0
setproctitle==1.3.2
tomli==2.0.1
uvicorn==0.21.1
